src/requirements.txt
asgiref==3.8.1
coverage==7.8.0
Django==5.2
djangorestframework==3.16.0
dotenv==0.9.9
flake8==7.2.0
gunicorn==23.0.0
iniconfig==2.1.0
mccabe==0.7.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pyflakes==3.3.2
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.1.0
sqlparse==0.5.3
whitenoise==6.9.0
